// src/generated/coc-api/api.ts
export interface BadgeUrls {
    'small': string;
    'medium': string;
    'large': string;
}
// ...
export interface IconUrls {
    'small'?: string | null;
    'medium'?: string | null;
    'tiny'?: string | null;
}
export interface Label {
    'name'?: string;
    'id'?: number;
    'iconUrls'?: IconUrls;
}
export interface LabelsObject {
    'items'?: Array<Label>;
}
export interface Language {
    'name'?: string;
    'id'?: number;
    'languageCode'?: string;
}
export interface League {
    'name'?: string;
    'id'?: number;
    'iconUrls'?: IconUrls;
}
export interface LeagueList {
    'items'?: Array<League>;
}
export interface LeagueSeason {
    'id'?: string;
}
export interface LeagueSeasonList {
    'items'?: Array<LeagueSeason>;
}
export interface LegendLeagueTournamentSeasonResult {
    'trophies'?: number;
    'id'?: string;
    'rank'?: number;
}
export interface Location {
    'localizedName'?: string | null;
    'id'?: number;
    'name'?: string;
    'isCountry'?: boolean;
    'countryCode'?: string | null;
}
export interface LocationList {
    'items'?: Array<Location>;
}
export interface Player {
    'clan'?: PlayerClan | null;
    'league'?: League | null;
    'role'?: Role | null;
    'warPreference'?: WarPreference | null;
    'attackWins'?: number;
    'defenseWins'?: number;
    'townHallLevel'?: number;
    'townHallWeaponLevel'?: number;
    'versusBattleWins'?: number;
    'legendStatistics'?: PlayerLegendStatistics | null;
    'troops'?: Array<PlayerItemLevel>;
    'heroes'?: Array<PlayerItemLevel>;
    'spells'?: Array<PlayerItemLevel>;
    'labels'?: Array<Label>;
    'tag'?: string;
    'name'?: string;
    'expLevel'?: number;
    'trophies'?: number;
    'bestTrophies'?: number;
    'donations'?: number;
    'donationsReceived'?: number;
    'builderHallLevel'?: number;
    'versusTrophies'?: number;
    'bestVersusTrophies'?: number;
    'warStars'?: number;
    'achievements'?: Array<PlayerAchievementProgress>;
    'versusBattleWinCount'?: number;
}


export interface PlayerAchievementProgress {
    'stars'?: number;
    'value'?: number;
    'name'?: string;
    'target'?: number;
    'info'?: string;
    'completionInfo'?: string | null;
    'village'?: string;
}
export interface PlayerClan {
    'tag'?: string;
    'clanLevel'?: number;
    'name'?: string;
    'badgeUrls'?: BadgeUrls;
}
export interface PlayerItemLevel {
    'level'?: number;
    'name'?: string;
    'maxLevel'?: number;
    'village'?: string;
    'superTroopIsActive'?: boolean | null;
}
export interface PlayerLegendStatistics {
    'currentSeason'?: LegendLeagueTournamentSeasonResult;
    'previousVersusSeason'?: LegendLeagueTournamentSeasonResult | null;
    'bestVersusSeason'?: LegendLeagueTournamentSeasonResult | null;
    'legendTrophies'?: number;
    'previousSeason'?: LegendLeagueTournamentSeasonResult | null;
    'bestSeason'?: LegendLeagueTournamentSeasonResult | null;
}
// ...
export const Role = {
    Member: 'member',
    Admin: 'admin',
    CoLeader: 'coLeader',
    Leader: 'leader'
} as const;

export type Role = typeof Role[keyof typeof Role];


export interface VerifyTokenRequest {
    'token'?: string;
}
// ...
export const WarPreference = {
    Out: 'out',
    In: 'in'
} as const;

export type WarPreference = typeof WarPreference[keyof typeof WarPreference];
// ...
export class PlayersApi extends BaseAPI {
    /**
     * Get information about a single player by player tag. Player tags can be found either in game or by from clan member lists. Note that player tags start with hash character \'#\' and that needs to be URL-encoded properly to work in URL, so for example player tag \'#2ABC\' would become \'%232ABC\' in the URL. 
     * @summary Get player information
     * @param {string} playerTag Tag of the player.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getPlayer(playerTag: string, options?: RawAxiosRequestConfig) {
        return PlayersApiFp(this.configuration).getPlayer(playerTag, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Verify player API token that can be found from the game settings. This API call can be used to check that players own the game accounts they claim to own as they need to provide the one-time use API token that exists inside the game. 
     * @summary Verify player API token that can be found from the game settings.
     * @param {string} playerTag Tag of the player.
     * @param {VerifyTokenRequest} body Request body
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public verifyToken(playerTag: string, body: VerifyTokenRequest, options?: RawAxiosRequestConfig) {
        return PlayersApiFp(this.configuration).verifyToken(playerTag, body, options).then((request) => request(this.axios, this.basePath));
    }
}